修复--前后端代码调试通过

+++ zhifan-server/src/main/java/com/zhifan/service/FileConfigService.java
@@ -16,6 +16,6 @@ public interface FileConfigService extends IService<FileConfig> {
 
     void monitorPath(String path);
 
-    void editOne(FileConfigReq req,int id);
+    void editOne(FileConfigReq req,Long id);
 
 }
